+++ src/main/java/org/folio/uk/configuration/SystemUserRoleRetryConfiguration.java
@@ -1,5 +1,6 @@
-package org.folio.uk.integration.kafka.configuration;
+package org.folio.uk.configuration;
 
+import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import java.time.Duration;
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Component;
 @Data
 @Component
 @ConfigurationProperties("application.retry.system-user-role")
-public class SystemUserRoleRetryConfiguration {
+public class SystemUserRoleRetryConfiguration implements RetryProperties {
 
   /**
    * Retry delay for module user capabilities assignment.
@@ -23,5 +24,6 @@ public class SystemUserRoleRetryConfiguration {
    */
   @NotNull
   @Positive
-  private int retryAttempts = 60;
+  @Max(Integer.MAX_VALUE)
+  private long retryAttempts = 60;
 }